@@ -5,44 +5,45 @@ import pipeline_bench
 
 
 def run_pipelines(pipeline_id, task_id, worker_dir):
-    try:
-        benchmark = pipeline_bench.Benchmark(
-            task_id=task_id,
-            worker_dir=worker_dir,
-            mode="live",
-        )
+    # try:
+    benchmark = pipeline_bench.Benchmark(
+        task_id=task_id,
+        worker_dir=worker_dir,
+        mode="live",
+    )
 
-        _ = benchmark(
-            pipeline_id=pipeline_id,
-        )
-    except Exception as e:
-        print(f"Failed for pipeline_id: {pipeline_id}")
-        print(e)
-        print()
+    _ = benchmark(
+        pipeline_id=pipeline_id,
+    )
+    # except Exception as e:
+    #     print(f"Failed for pipeline_id: {pipeline_id}")
+    #     print(e)
+    #     print()
 
 
 def process_pipeline(pipeline_id, task_id, worker_dir):
-    p = multiprocessing.Process(
-        target=run_pipelines, args=(pipeline_id, task_id, worker_dir)
-    )
-    p.start()
+    run_pipelines(pipeline_id, task_id, worker_dir)
+    # p = multiprocessing.Process(
+    #     target=run_pipelines, args=(pipeline_id, task_id, worker_dir)
+    # )
+    # p.start()
 
-    # Wait for 1200 seconds or until process finishes
-    p.join(2400)
+    # # Wait for 1200 seconds or until process finishes
+    # p.join(2400)
 
-    # If thread is still active
-    if p.is_alive():
-        print("Running... let's kill it...")
+    # # If thread is still active
+    # if p.is_alive():
+    #     print("Running... let's kill it...")
 
-        # Terminate - may not work if process catches the signal and doesn't exit
-        p.terminate()
+    #     # Terminate - may not work if process catches the signal and doesn't exit
+    #     p.terminate()
 
-        # Check if process has really terminated & print message
-        time.sleep(0.1)
-        if p.is_alive():
-            print("Failed to terminate the process")
-        else:
-            print(f"Process terminated for pipeline_id: {pipeline_id}")
+    #     # Check if process has really terminated & print message
+    #     time.sleep(0.1)
+    #     if p.is_alive():
+    #         print("Failed to terminate the process")
+    #     else:
+    #         print(f"Process terminated for pipeline_id: {pipeline_id}")
 
 
 if __name__ == "__main__":
@@ -51,7 +52,7 @@ if __name__ == "__main__":
     parser.add_argument("--end_pipeline_id", type=int, required=True)
     parser.add_argument("--task_id", type=int, required=True)
     parser.add_argument(
-        "--worker_dir", type=str, default="/work/ws/nemo/fr_mj237-pipeline_bench-0/"
+        "--worker_dir", type=str, default="/work/dlclarge1/janowski-pipebench"
     )
     args = parser.parse_args()
 
